+++ config/odometer-easy.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Gsferro/FilamentOdometerEasy
-return [
-
-];