+++ src/Facades/FilamentOdometerEasy.php
@@ -3,8 +3,13 @@
 namespace Gsferro\FilamentOdometerEasy\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Illuminate\Support\HtmlString;
 
 /**
+ * @method static HtmlString render(mixed $value, string|array|null $format = null, ?string $class = null, ?string $driver = null)
+ * @method static HtmlString renderNumberFlow(mixed $value, ?array $format = null, ?string $class = null)
+ * @method static HtmlString renderOdometer(mixed $value, ?string $format = null, ?string $class = null)
+ *
  * @see \Gsferro\FilamentOdometerEasy\FilamentOdometerEasy
  */
 class FilamentOdometerEasy extends Facade